requirements.txt
confluent-kafka==2.8.2
web3==7.13.0
websockets==15.0.1
pydantic==2.11.4
PyYAML==6.0.2
